code refactored and game guide button implemented to assist user with game instructions
