+++ src/ants/ScubaThrowerAnt.java
@@ -1,6 +1,5 @@
 package ants;
 
-import core.ScubaAnt;
 import core.AntColony;
 import core.Bee;
