+++ src/core/AntsVsSomeBees.java
@@ -11,7 +11,7 @@ public class AntsVsSomeBees
 	public static int startingFood = 25; // ==Niall== I just added this so that it could be easily identified and altered.
 
 	public static void main(String[] args) throws IOException, UnsupportedAudioFileException, LineUnavailableException {
-		AntColony colony = new AntColony(3, 8, 0, startingFood); //specify the colony [tunnels, length, moats, food]
+		AntColony colony = new AntColony(3, 8,0, startingFood); //specify the colony [tunnels, length, moats, food]
 		//Hive hive = Hive.makeTestHive(); //specify the attackers (the hive)
 		Hive hive = Hive.makeFullHive();
 		//Hive hive = Hive.makeInsaneHive();
